Extract mobile host and browser helpers
