+++ mobile/src/host-screen/host-screen-styles.ts
@@ -1,0 +1,7 @@
+import { hostScreenPrimaryStyles } from './host-screen-primary-styles'
+import { hostScreenSecondaryStyles } from './host-screen-secondary-styles'
+
+export const hostScreenStyles = {
+  ...hostScreenPrimaryStyles,
+  ...hostScreenSecondaryStyles
+}
